+++ CHANGELOG.md
@@ -9,7 +9,13 @@ v2.11 (Anoia) - (in development)
     - Tools > Registration > Move bounding-box center to origin
     - Tools > Registration > Move bounding-box min corner to origin
     - Tools > Registration > Move bounding-box max corner to origin
+
 - Improvements
+  - Clipping box tool:
+    - option to select the extracted contour type (LOWER, UPPER or FULL)
+	  - The 'up' direction is always Z for slices normal to X or Y (the local X, Y or Z directions of the active cross-section)
+	  - The 'up' direction is X for slices normal to Z
+	- all parameters should now be properly remembered from one call to the other (during the same session)
   - Command line tool:
     - The C2M_DIST command (Cloud-to-Mesh distances) can now be called with 2 meshes as input.
         In this case the first mesh vertices are used as compared cloud.